[FIX] website_sale: checkout accordion card styling
The goal of this PR is to unify the styling of the accordion component
with the bootstrap styling of a card.

The layout of the accordion had to be adapted with an extra div to use
the `accordion-item` as the `card-body`. This styling will allow better
maintainability of the accordion which now visually act as a card
component in > lg screens. This will be required to handle different
background theme color.

This PR also fixes wrong indentation in the SCSS.

task-3556135

closes odoo/odoo#138522

Signed-off-by: Valentin Chevalier <vcr@odoo.com>